proposal: add initial implementation of actions and history
This changeset adds the concept of a generic log event which can be
attached to any task, and displayed in the UI. The action handlers for
the log events also trigger events, so "log event" may be the wrong
word to apply to the controllers.

The layout for document submission tasks is still in flux, but I don't
expect it to change too much. Adding a column on the right-hand side
makes the instructions _much_ more readable.

Remaining TODOs:

* Authorization
* Display validation messages properly for actions (this is harder than it
  sounds, since we need to re-render an arbitrary task)
* Beautification
* A javascript control to toggle the visibility of the instructions
* Some notion of role and the current state